Remove misleading "deep equality" claim.
Testing tuple equality with === gives results identical to testing the
*shallow* identities of the elements in the tuple. Equality checking does
not recurse further into the elements of the tuple (nor should it), so
"deep equality checking" could be misleading.

Deeply nested tuples containing other tuples can be deeply compared in
constant time using ===, of course, and this property would extend to
other internalized persistent data structures. Nevertheless, it's
important for users to realize that this library treats "normal" objects
and primitive values the same way a Map would treat them.

With that said, "constant-time equality testing (===)" is still an
accurate claim without the "deep" adjective.